bot-controlled: move webhooks code around and refactor arg parsing
